src/backend/requirements.txt
crewai>=0.11.0
langchain>=0.1.0
langchain-community>=0.0.1
python-dotenv>=0.19.0
websockets>=10.0
aiofiles>=23.0
pyyaml>=6.0.0
openai>=1.0.0  # Required for CrewAI's default LLM support
google-cloud-aiplatform>=1.35.0  # For VertexAI
anthropic>=0.7.0  # For Claude models
google-generativeai>=0.3.0  # For Gemini
openrouter>=0.3.0  # For OpenRouter access
botocore>=1.31.17  # Required for some provider authentications
requests>=2.31.0  # Required for API calls
aiohttp>=3.9.0  # For async operations
httpx>=0.24.1  # For modern HTTP requests
pydantic>=2.0.0  # For data validation
